+++ src/ngrok/server/cli.go
@@ -16,12 +16,12 @@ type Options struct {
 }
 
 func parseArgs() *Options {
-	httpAddr := flag.String("httpAddr", ":80", "Public address for HTTP connections, empty string to disable")
-	httpsAddr := flag.String("httpsAddr", ":443", "Public address listening for HTTPS connections, emptry string to disable")
-	tunnelAddr := flag.String("tunnelAddr", ":4443", "Public address listening for ngrok client")
+	httpAddr := flag.String("http-addr", ":80", "Public address for HTTP connections, empty string to disable")
+	httpsAddr := flag.String("https-addr", ":443", "Public address listening for HTTPS connections, emptry string to disable")
+	tunnelAddr := flag.String("tunnel-addr", ":4443", "Public address listening for ngrok client")
 	domain := flag.String("domain", "ngrok.com", "Domain where the tunnels are hosted")
-	tlsCrt := flag.String("tlsCrt", "", "Path to a TLS certificate file")
-	tlsKey := flag.String("tlsKey", "", "Path to a TLS key file")
+	tlsCrt := flag.String("tls-cert", "", "Path to a TLS certificate file, default path '/etc/ngrok/tls/ngrok.me/server.crt'")
+	tlsKey := flag.String("tls-key", "", "Path to a TLS key file, default path '/etc/ngrok/tls/ngrok.me/server.key'")
 	logto := flag.String("log", "stdout", "Write log messages to this file. 'stdout' and 'none' have special meanings")
 	loglevel := flag.String("log-level", "DEBUG", "The level of messages to log. One of: DEBUG, INFO, WARNING, ERROR")
 	flag.Parse()