+++ src/ngrok/server/tls.go
@@ -3,18 +3,19 @@ package server
 import (
 	"crypto/tls"
 	"io/ioutil"
-	"ngrok/server/assets"
+)
+
+const (
+	defaultTlsCrtPath = "/etc/ngrok/tls/ngrok.me/server.crt"
+	defaultTlsKeyPath = "/etc/ngrok/tls/ngrok.me/server.key"
 )
 
 func LoadTLSConfig(crtPath string, keyPath string) (tlsConfig *tls.Config, err error) {
-	fileOrAsset := func(path string, default_path string) ([]byte, error) {
-		loadFn := ioutil.ReadFile
+	fileOrAsset := func(path, defaultPath string) ([]byte, error) {
 		if path == "" {
-			loadFn = assets.Asset
-			path = default_path
+			return ioutil.ReadFile(defaultPath)
 		}
-
-		return loadFn(path)
+		return ioutil.ReadFile(path)
 	}
 
 	var (
@@ -23,11 +24,11 @@ func LoadTLSConfig(crtPath string, keyPath string) (tlsConfig *tls.Config, err e
 		cert tls.Certificate
 	)
 
-	if crt, err = fileOrAsset(crtPath, "assets/server/tls/snakeoil.crt"); err != nil {
+	if crt, err = fileOrAsset(crtPath, defaultTlsCrtPath); err != nil {
 		return
 	}
 
-	if key, err = fileOrAsset(keyPath, "assets/server/tls/snakeoil.key"); err != nil {
+	if key, err = fileOrAsset(keyPath, defaultTlsKeyPath); err != nil {
 		return
 	}